`common.sh`: execute `copy_to_docker` only when in docker mode

+++ packages/contracts-bedrock/test/kontrol/scripts/common.sh
@@ -141,13 +141,15 @@ copy_to_docker() {
   #   Error response from daemon: invalid symlink "/home/user/workspace/node_modules/@typescript-eslint/eslint-plugin" -> "../../../../node_modules/.pnpm/@typescript-eslint+eslint-plugin@6.19.1_@typescript-eslint+parser@6.19.1_eslint@8.56.0_typescript@5.3.3/node_modules/@typescript-eslint/eslint-plugin"
   # Even though we use a bind mount, we still need to copy the files to the
   # container because we are running Docker on a remote host.
-  TMP_DIR=$(mktemp -d)
-  cp -r "$WORKSPACE_DIR/." "$TMP_DIR"
-  rm -rf "$TMP_DIR/node_modules"
-  docker cp --follow-link "$TMP_DIR/." $CONTAINER_NAME:/home/user/workspace
-  rm -rf "$TMP_DIR"
+  if [ "$LOCAL" == false ]; then
+    TMP_DIR=$(mktemp -d)
+    cp -r "$WORKSPACE_DIR/." "$TMP_DIR"
+    rm -rf "$TMP_DIR/node_modules"
+    docker cp --follow-link "$TMP_DIR/." $CONTAINER_NAME:/home/user/workspace
+    rm -rf "$TMP_DIR"
 
-  docker exec --user root "$CONTAINER_NAME" chown -R user:user /home/user
+    docker exec --user root "$CONTAINER_NAME" chown -R user:user /home/user
+  fi
 }
 
 clean_docker(){
